STAC-11697 Fixes bug where copy_images.sh would fail for image specs without a registry prefix like docker.io/
Also:
- Updates helm template command for 4.2.x version of Helm chart
- Adds -t flag to perform a dry-run of the copy
- Adds -c and -r flags to set chart and repository to use

+++ stable/stackstate/installation/copy_images.sh
@@ -4,9 +4,12 @@ set -e
 
 # global
 is_ecr_re='\.ecr\..*\.amazonaws\.com'
-repo_and_tag_re='/([^/]*/[^/:]*):([^/:]*)$'
+repo_and_tag_re='^([^/:]+/)?([^/:]+/[^/:]+):([^/:]+)$'
 nc="\033[0m"
 red="\\033[0;31m"
+helm_chart=stackstate
+helm_repository=https://helm.stackstate.io
+dry_run=false
 
 # usage
 function usage() {
@@ -14,16 +17,23 @@ function usage() {
 Copy Docker images needed for StackState chart to another Docker image registry
 
 Arguments:
+    -c : Helm chart (default: $helm_chart)
     -d : Destination Docker image registry (required)
     -h : Show this help text
+    -r : Helm repository (default: $helm_repository)
+    -t : Dry-run
 EOF
 }
 
+
 # parse options
-while getopts "d:h" opt; do
+while getopts "c:d:hr:t" opt; do
   case "$opt" in
+  c) helm_chart=$OPTARG ;;
   d) dest_registry=$OPTARG ;;
-  h) usage; exit;;
+  h) usage; exit ;;
+  r) helm_repository=$OPTARG ;;
+  t) dry_run=true ;;
   \?) echo "Unknown option: -$OPTARG" >&2; exit 1;;
   :) echo "Missing option argument for -$OPTARG" >&2; exit 1;;
   *) echo "Unimplemented option: -$OPTARG" >&2; exit 1;;
@@ -31,26 +41,31 @@ while getopts "d:h" opt; do
 done
 shift $((OPTIND -1))
 
-[ -z "$dest_registry" ] && echo -e "${red}Provide the destination registry with the -d flag${nc}" && exit 1
+[ -z "$dest_registry" ] && echo -e "${red}Provide the destination registry with the -d flag${nc}" && usage && exit 1
+
 
 #
 images=()
-while IFS='' read -r line; do images+=("$line"); done < <(helm template stackstate stackstate --repo https://helm.stackstate.io --set stackstate.license.key=dummy | grep image: | sed -E 's/^.*image: "?([^"]*)"?.*$/\1/')
+while IFS='' read -r line; do images+=("$line"); done < <(helm template stackstate "$helm_chart" --repo "$helm_repository" --set "stackstate.baseUrl=http://dummy,stackstate.admin.authentication.password=dummy,stackstate.authentication.adminPassword=dummy,stackstate.license.key=dummy" | grep image: | sed -E 's/^.*image: "?([^"]*)"?.*$/\1/')
 for src_image in "${images[@]}"
 do
     if [[ "$src_image" =~ $repo_and_tag_re ]]; then
-        repo=${BASH_REMATCH[1]}
-        tag=${BASH_REMATCH[2]}
+        repo=${BASH_REMATCH[2]}
+        tag=${BASH_REMATCH[3]}
         dest_image="$dest_registry/$repo:$tag"
-        if [[ "$dest_registry" =~ $is_ecr_re ]]; then
-            echo "Ensuring ECR repository $repo exists"
-            aws ecr describe-repositories --repository-names "$repo" >/dev/null 2>/dev/null || aws ecr create-repository --repository-name "$repo" > /dev/null
+        if $dry_run; then
+            echo "Copying $src_image to $dest_image (dry-run)"
+        else
+            if [[ "$dest_registry" =~ $is_ecr_re ]]; then
+                echo "Ensuring ECR repository $repo exists"
+                aws ecr describe-repositories --repository-names "$repo" >/dev/null 2>/dev/null || aws ecr create-repository --repository-name "$repo" > /dev/null
+            fi
+            echo "Copying $src_image to $dest_image"
+            docker pull "$src_image"
+            docker tag "$src_image" "$dest_image"
+            docker push "$dest_image"
+            docker rmi "$dest_image"
         fi
-        echo "Copying $src_image to $dest_image"
-        docker pull "$src_image"
-        docker tag "$src_image" "$dest_image"
-        docker push "$dest_image"
-        docker rmi "$dest_image"
     else
         1>&2 echo "Cannot determine repository and tag for $src_image"
         exit 1
